% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyr_rename_for_openair.R
\name{hyr_rename_for_openair}
\alias{hyr_rename_for_openair}
\title{Function to rename variables in a tibble created by \code{\link{read_hyr}} to
enable immediate use with \strong{openair}.}
\usage{
hyr_rename_for_openair(df)
}
\arguments{
\item{df}{Tibble from \code{\link{read_hyr}}.}
}
\value{
Tibble.
}
\description{
Function to rename variables in a tibble created by \code{\link{read_hyr}} to
enable immediate use with \strong{openair}.
}
\seealso{
\code{\link{read_hyr}}
}
\author{
Stuart K. Grange
}
