% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_package_exports.R
\name{lubridate-functions}
\alias{lubridate-functions}
\title{Pseudo-function to re-export \strong{lubridate}'s common functions.}
\description{
Pseudo-function to re-export \strong{lubridate}'s common functions.
}
