% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hyr.R
\name{read_hyr}
\alias{read_hyr}
\title{Function to load HYSPLIT data files produced with \code{\link{hyr_run}}.}
\usage{
read_hyr(file, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to load HYSPLIT data files produced with \code{\link{hyr_run}}.
}
\author{
Stuart K. Grange
}
