% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hysplit.R
\name{run_hysplit}
\alias{run_hysplit}
\title{Function to command HYSPLIT (Hybrid Single Particle Lagrangian Integrated 
Trajectory Model) from R.}
\usage{
run_hysplit(latitude = 51.5, longitude = -0.1, start = NA, end = NA,
  interval = "3 hour", runtime = -96, start_height = 10,
  model_height = 10000, hysplit_exec = "exec/", hysplit_input,
  hysplit_output = "hysplit_output/", delete = TRUE, drop = TRUE,
  site = NA, source = NA, verbose = FALSE)
}
\arguments{
\item{latitude}{Latitude of a receptor location.}

\item{longitude}{Longitude of a receptor location.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{interval}{What interval should the trajectories be run at? Default is 
\code{"3 hour"}, as in \strong{openair}.}

\item{runtime}{How many hours from time zero should the trajectory be run? 
Default is \code{-96}, i.e., a back trajectory of 96 hours, as in 
\strong{openair}.}

\item{start_height}{Height of receptor location. Default is \code{10} metres, 
as in \strong{openair}.}

\item{model_height}{Height of model domain. Default is \code{10000} metres,
as in \strong{openair}.}

\item{hysplit_exec}{Location of HYSPLIT's executable files. Ensure executable
permisions have been set for the \code{hyts_std} application if on a Unix 
system. HYSPLIT's \code{bdyfiles} directory with the \code{ASCDATA.CFG} will
also need to be in the correct location.}

\item{hysplit_input}{Location of input meteorological files for HYSPLIT.}

\item{hysplit_output}{Location of where HYSPLIT should write its trajectory 
outputs while working.}

\item{delete}{Should temporary files be silently deleted before and after the
model runs? Default is \code{TRUE}.}

\item{drop}{Should the \code{"year"}, \code{"month"}, \code{"day"}, 
\code{"hour"}, and \code{"receptor"} variables be dropped from the bound
object? Default is \code{TRUE}.}

\item{site}{An optional site string to be added to the returned data frame.}

\item{source}{An optional source string to be added to the returned data frame.}

\item{verbose}{Should the function give messages on what trajectory is being
processed. Default is \code{FALSE} and will estimate the time until 
completion.}
}
\description{
\code{run_hysplit} will return a data frame which is ready for use in 
\strong{openair}'s \code{traj*} functions.
}
\examples{
\dontrun{

# Run back trajectories for 2010 for Lerwick, Shetland Islands
data_lerwick <- run_hysplit(
  latitude = 60.13922, 
  longitude = -1.185319, 
  runtime = -96, 
  start_height = 10, 
  model_height = 10000, 
  start = 2010,
  end = 2010,
  hysplit_exec = "/AQprojects/hysplit/exec", 
  hysplit_input = "/AQprojects/Hysplit/TrajData", 
  hysplit_output = "/AQprojects/hysplit/hysplit_output",
  site = "lerw"
)
  
# Run back trajectories for 2015 at 1500 metres and extended run time for
# analysis of air aloft or Gibraltar
data_gibraltar <- run_hysplit(
  latitude = 36.134, 
  longitude = -5.347, 
  runtime = -240, 
  start_height = 1500, 
  model_height = 10000, 
  start = 2015,
  end = 2015,
  hysplit_exec = "~/Dropbox/ricardo/hysplit/exec", 
  hysplit_input = "/media/stuartg/MOBILE_II/data_objects/hysplit_met_data", 
  hysplit_output = "~/Dropbox/ricardo/hysplit/hysplit_output",
  site = "gibraltar"
)
  
# On a Windows system with a default hysplit installation
data_gibraltar_standard <- run_hysplit(
  latitude = 36.134, 
  longitude = -5.347, 
  runtime = -240, 
  start_height = 10, 
  model_height = 10000, 
  start = 2015,
  end = 2015,
  hysplit_exec = "C:/hysplit4/exec", 
  hysplit_input = "D:/data/hysplit_met_data", 
  hysplit_output = "C:/hysplit4/output",
  site = "gibraltar"
)
  
}

}
\seealso{
\code{\link{importTraj}}, \code{\link{trajPlot}}, 
\href{http://ready.arl.noaa.gov/HYSPLIT.php}{HYSPLIT home}
}
\author{
Stuart K. Grange
}
