% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyr_rename_for_openair.R
\name{hyr_rename_for_openair}
\alias{hyr_rename_for_openair}
\title{Function to rename variables in a data frame containing trajectory data for
use in \strong{openair}.}
\usage{
hyr_rename_for_openair(df)
}
\arguments{
\item{df}{Data frame from \code{\link{read_hyr}}.}
}
\value{
Tibble.
}
\description{
Function to rename variables in a data frame containing trajectory data for
use in \strong{openair}.
}
\author{
Stuart K. Grange
}
