% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyr_run.R
\name{hyr_run}
\alias{hyr_run}
\title{Function to command HYSPLIT (Hybrid Single Particle Lagrangian Integrated 
Trajectory Model) from R.}
\usage{
hyr_run(df, directory_exec = "exec/", directory_input, directory_output,
  verbose = TRUE)
}
\arguments{
\item{df}{Data}

\item{directory_exec}{Location of HYSPLIT's executable files. Ensure executable
permisions have been set for the \code{hyts_std} application if on a Unix 
system. HYSPLIT's \code{bdyfiles} directory with the \code{ASCDATA.CFG} will
also need to be in the correct location.}

\item{directory_input}{Location of input meteorological files for HYSPLIT.}

\item{directory_output}{Location of where HYSPLIT should write its trajectory 
outputs while working.}

\item{verbose}{Should the function give messages on what trajectory is being
processed.}
}
\description{
\code{run_hysplit} will return a data frame which is ready for use in 
\strong{openair}'s \code{traj*} functions.
}
\seealso{
\code{\link[openair]{importTraj}}, \code{\link[openair]{trajPlot}}, 
\href{http://ready.arl.noaa.gov/HYSPLIT.php}{HYSPLIT home}
}
\author{
Stuart K. Grange
}
